+++ dart/geobase/CHANGELOG.md
@@ -6,6 +6,7 @@ NOTE: Version 1.1.0 currently under development (1.1.0-dev.0).
 
 🧩 Features:
 * [Enhanced coordinate value output for positions and text data formats #98](https://github.com/navibyte/geospatial/issues/98)
+* [Add decodeHex and toBytesHex to geometry objects and builders #227](https://github.com/navibyte/geospatial/issues/227)
 
 ## 1.0.2